Point::operator==() now just access the other point's value vector directly now, instead of using its own [] operator. This causes less function call overhead if the method was not inlined.
std::fabs() is now being used instead of the C fabs(). fabs() is meant for doubles, but floats are being used for this system. Therefore, two casts were performed for every compare() call, which had a surprisingly high impact on performance. std::fabs() avoids this cast because is has an explicit overload for floats.

There was another case where a vector was assigned BY VALUE, meaning its contents were copied. This was when an index is inserted into an already existing bucket. This has been solved by directly adding the index to a REFERENCE to the index list.